Changing from 64 bits to 32 bits

+++ test/runtests.jl
@@ -1,8 +1,8 @@
 using Multifractal 
 
-function main(inputfile::ASCIIString, extensionDq::ASCIIString, extensionFa::ASCIIString, initialQ::Float64, finalQ::Float64, dq::Float64, Np::Int32, r2dq::Float64, r2fa::Float64, scalesToRemove::Int32)
+function main(inputfile::ASCIIString, extensionDq::ASCIIString, extensionFa::ASCIIString, initialQ::Float32, finalQ::Float32, dq::Float32, Np::Int32, r2dq::Float32, r2fa::Float32, scalesToRemove::Int32)
     Multifractal.ChhabraJansen(inputfile, extensionDq, extensionFa, initialQ, finalQ, dq, Np, r2dq, r2fa, scalesToRemove)
 end
 
-#main(ARGS[1], ARGS[2], ARGS[3], parse(Float64,ARGS[4]), parse(Float64,ARGS[5]), parse(Float64,ARGS[6]), parse(Int,ARGS[7]), parse(Float64,ARGS[8]), parse(Float64,ARGS[9]),parse(Int,ARGS[10]));
+#main(ARGS[1], ARGS[2], ARGS[3], parse(Float32,ARGS[4]), parse(Float32,ARGS[5]), parse(Float32,ARGS[6]), parse(Int,ARGS[7]), parse(Float32,ARGS[8]), parse(Float32,ARGS[9]),parse(Int,ARGS[10]));
 main("series.txt","tdq","tfa",-5.0,5.0,1.0,9,-1.0,-1.0,0);
